use int64

+++ schema/article.go
@@ -5,8 +5,8 @@ import "time"
 // Article Article schema
 //+migu table:"articles"
 type Article struct {
-	ID        int `migu:"pk,autoincrement"`
-	UserID    int
+	ID        int64 `migu:"pk,autoincrement"`
+	UserID    int64
 	Text      string    `migu:"type:text"`
 	CreatedAt time.Time `migu:"default:CURRENT_TIMESTAMP"`
 	UpdatedAt time.Time `migu:"default:CURRENT_TIMESTAMP,extra:ON UPDATE CURRENT_TIMESTAMP"`
